Use Jasmine method to set default timeout
Instead of sending an explicit timeout of 10 seconds on each call, use
Jasmine's own method to set the default timeout to 10 seconds. This
allows specs of projects using `jasmine-fix` to use the same method to
set the default timeout globally in one place instead of having to
explicitly send in their own override timeout.